Smaller primitives in minutiae and edges

+++ src/main/java/com/machinezoo/sourceafis/transparency/types/EdgeShape.java
@@ -4,7 +4,7 @@ package com.machinezoo.sourceafis.transparency.types;
 import java.io.*;
 
 public interface EdgeShape extends Serializable {
-	int length();
-	@Angle double referenceAngle();
-	@Angle double neighborAngle();
+	short length();
+	@Angle float referenceAngle();
+	@Angle float neighborAngle();
 }
